Add option to disable multi results
https://github.com/github/trilogy/pull/57 set multi-results as a
default.

To safely upgrade to the latest version of Trilogy, we want to turn that
off for now, and then turn it on one cluster at a time.

This commit removes the default from the C library and into the Ruby C
extension so we can disable it as needed. (We had an option to enable
multi results, but removed it in
https://github.com/github/trilogy/pull/68 because the option was always
a no-op. This PR adds the option back, with some modification).